Update interview-packet.md
Change the job spec to sound as though it is coming from the team instead of James.  Removed the What We Are Looking For section because that is redundant with what is the job specification.  Removed the link to the job specification because it is no longer live.  Added Dan Spiegel under engineering.